Refatora a estrutura de cache, movendo a lógica para a classe RedisCache e criando a interface Cache
